Data for Diallel Analysis using Griffing Approach Method 1

Griffing is used for performing Diallel Analysis using Griffing's Approach.

data(GriffingData1)

Format

A data.frame with 256 rows and 4 variables.

Details

Cross1 Cross 1

Cross2 Cross 2

Rep Replicate

Yield Yield Response

References

Griffing, B. (1956) Concept of General and Specific Combining Ability in relation to Diallel Crossing Systems. Australian Journal of Biological Sciences, 9(4), 463--493.

Singh, R. K. and Chaudhary, B. D. (2004) Biometrical Methods in Quantitative Genetic Analysis. New Delhi: Kalyani.

See also

Examples

data(GriffingData1)